restructured API route
